Clubstations for Wavelog (#1334)
* feat[clubstations]: New DB structure

* feat[clubstations]: Add clubstationstable in user managment

* feat[clubstations]: Show last operator

* feat[clubstations]: Better solution for last operator. tnx for the hint @int2001

* feat[clubstations]: New Club Model and Controller

* feat[clubstations]: Add "Add User" and "Edit User" functionality

* docs[clubstations]: move comment

* feat[clubstations]: Add "Delete Member" functionality

* feat[clubstations]: some enhancements and javascript

* fix[clubstations]: Wrong message class for flashmessages

* feat[clubstations]: Added Switch in the Header menu (not functional yet)

* feat[clubstations]: clubswitch modal

* fix[clubstations]: Load encryption library if not already loaded

* fix[clubstations]: Prevent direct login attempts to clubstations and enhance impersonation authorization

* fix[clubstations]: Typo

* feat[clubstations]: Only show the operator dialog if there is something fishy

* fix[user]: little UI bug

* feat[impersonate]: Add source uid to session data

* fix[impersonate]: logic adjustment

* feat[clubstations]: Add manage button in header menu for club officers

* fix[clubstations]: typo in permission level check

* fix[clubstations]: Full rights for the admin

* feat[impersonate]: Custom sessiondata

* feat[impersonate]: Implement stop impersonation feature with modal confirmation; "the way back"

* fix(modal): Fix bug where modal was hidden when mouse leaved the browser content

* docs(config): Adjust config description for special callsigns and clubstations

* feat(club): Add club access check helper

* typo

* fix[impersonation]: Better text

* feat(club): Selectize for a efficient user search

* feat(clubstations): Restrict clubstations based on users permission level part 1/x

* adjustments for dev merge

* Adjusted club right for the advanced logbook

* feat[user]: Refactoring of the Action Buttons in the user table

* fix[club_permissions]: normal button instead small one for club permissions

* remove unnecessary line break in modal body

* feat[clubstations]: Add Club Mode badge to the header

* fix[clubstations]: fix maintenance mode

* allow switch back on http

* feat(simplefle): display operator input based on club_access

* small UI adjustments

* small UI adjustments

* moved api page to a index.php file and added support for clubstations

* removed unused stuff

* typo

* radios and api keys

* missed one binding

* fix qso view, even officers do just see their own radios in QSO logging

* omit the need for a relogin to see the changes as an admin

* Omit the need for relogin after club changes in general. It's a question of UX. It's better to accept a little higher DB load (if clubstations are enabled) then the need of an user to relogin. There is some room for improvement by changing user_model->get_by_id() and adding a join there. This can be done later if we see that the load is too high

* If the user is not the creator of the API key, it's likely a clubstation. In this case the callsign of the clubstation can not be the same as the callsign of the user (operator call provided by the user). If this is the case, we need to use the callsign of the creator of the API key

* remove debug messages

* better UI in header

* found a typo

* full access in clubstations for admins (if accessed via admin usertable)

* adjusted text

* adjusted text

* adjust text

* reduce required chars

* bugfix: missing the correct authentication in case the admin was not member of the club. he wasn't able to switch back

* reduce debug messages

* fixed UI bug related to tooltips

* load js in controller

* upps..

* some UI adjustments

* corrected permissions

* if user gets delete we need to remove data in club_permissions and also api keys which were created by this user

* Notify members about new memberships or changes in permission level

* add spinner to save button

* make login/logout process more bulletproof

* remove the relogin cookie after the attempt

* better strategy

* bug where switch back failed if user is no admin

* make api keys more secure

* mask not owned api keys

* removed annoying link

* if a user gets removed from a club we also should delete the corresponding api keys and cat radios

* adjusted wiki link

* Auto creation of logbook and location when new user is created

* store and display locator in uppercase

* same for callsign

* fixed a bug in user/club creation

* Revert "Auto creation of logbook and location when new user is created"
We found another solution to which will be addressed in a second PR
This reverts commit f05f4b7bf0423a88abf0087ade81fe613d217794.

* Optimized SQL for stats at userlist

* Source query for lastop "out", because mysql<9.0 can't handle Windowed functions

* adjust migration

* add new columns to users table to get created_at and modified_at

* added a partial down function

* add operator dropdown for clubstations

* fix mig version

* Add some backend restrictions in case a user wants to try something funny with the club

---------

+++ application/models/User_model.php
@@ -154,6 +154,29 @@ class User_Model extends CI_Model {
 		}
 	}
 
+	// FUNCTION: array search_users($query)
+	// Search for users by parts of their callsign
+	function search_users($query, $clubstations = false) {
+		if (strlen($query) < 2) {
+			return false;
+		}
+		$this->db->select('user_id, user_callsign, user_firstname, user_lastname');
+		if (!$clubstations) {
+			$this->db->where('clubstation', 0);
+		}
+
+		$this->db->group_start();
+		$this->db->like('user_callsign', $query);
+		$this->db->or_like('user_firstname', $query);
+		$this->db->or_like('user_lastname', $query);
+		$this->db->group_end();
+
+		$this->db->limit(100);
+
+		$r = $this->db->get($this->config->item('auth_table'));
+		return $r;
+	}
+
 	// FUNCTION: bool add($username, $password, $email, $type)
 	// Add a user
 	function add($username, $password, $email, $type, $firstname, $lastname, $callsign, $locator, $timezone,
@@ -164,7 +187,7 @@ class User_Model extends CI_Model {
 		$user_language, $user_hamsat_key, $user_hamsat_workable_only, $user_iota_to_qso_tab, $user_sota_to_qso_tab,
 		$user_wwff_to_qso_tab, $user_pota_to_qso_tab, $user_sig_to_qso_tab, $user_dok_to_qso_tab,
 		$user_lotw_name, $user_lotw_password, $user_eqsl_name, $user_eqsl_password, $user_clublog_name, $user_clublog_password,
-		$user_winkey) {
+		$user_winkey, $clubstation = 0) {
 		// Check that the user isn't already used
 		if(!$this->exists($username)) {
 			$data = array(
@@ -172,10 +195,10 @@ class User_Model extends CI_Model {
 				'user_password' => $this->_hash($password),
 				'user_email' => xss_clean($email),
 				'user_type' => xss_clean($type),
-				'user_firstname' => xss_clean($firstname),
-				'user_lastname' => xss_clean($lastname),
-				'user_callsign' => xss_clean($callsign),
-				'user_locator' => xss_clean($locator),
+				'user_firstname' => xss_clean($firstname) ?? '',
+				'user_lastname' => xss_clean($lastname) ?? '',
+				'user_callsign' => strtoupper(xss_clean($callsign)),
+				'user_locator' => strtoupper(xss_clean($locator)),
 				'user_timezone' => xss_clean($timezone),
 				'user_measurement_base' => xss_clean($measurement),
 				'user_date_format' => xss_clean($user_date_format),
@@ -206,7 +229,8 @@ class User_Model extends CI_Model {
 				'user_eqsl_password' => xss_clean($user_eqsl_password),
 				'user_clublog_name' => xss_clean($user_clublog_name),
 				'user_clublog_password' => xss_clean($user_clublog_password),
-				'winkey' => xss_clean($user_winkey)
+				'winkey' => xss_clean($user_winkey),
+				'clubstation' => $clubstation,
 			);
 
 			// Check the password is valid
@@ -254,8 +278,8 @@ class User_Model extends CI_Model {
 				$data = array(
 					'user_name' => xss_clean($fields['user_name']),
 					'user_email' => xss_clean($fields['user_email']),
-					'user_callsign' => xss_clean($fields['user_callsign']),
-					'user_locator' => xss_clean($fields['user_locator']),
+					'user_callsign' => strtoupper(xss_clean($fields['user_callsign'])),
+					'user_locator' => strtoupper(xss_clean($fields['user_locator'])),
 					'user_firstname' => xss_clean($fields['user_firstname']),
 					'user_lastname' => xss_clean($fields['user_lastname']),
 					'user_timezone' => xss_clean($fields['user_timezone']),
@@ -388,7 +412,8 @@ class User_Model extends CI_Model {
 			}
 			// Delete QSOs from $this->config->item('table_name')
 			$this->db->query("DELETE FROM bandxuser WHERE userid = ?",$user_id);
-			$this->db->query("DELETE FROM api WHERE user_id = ?",$user_id);
+			$this->db->query("DELETE FROM api WHERE user_id = ? OR created_by = ?", [$user_id, $user_id]);
+			$this->db->query("DELETE FROM club_permissions WHERE user_id = ? OR club_id = ?", [$user_id, $user_id]);
 			$this->db->query("DELETE FROM cat WHERE user_id = ?",$user_id);
 			$this->db->query("DELETE FROM lotw_certs WHERE user_id = ?",$user_id);
 			$this->db->query("DELETE FROM notes WHERE user_id = ?",$user_id);
@@ -397,7 +422,7 @@ class User_Model extends CI_Model {
 			$this->db->query("DELETE FROM queries WHERE userid = ?",$user_id);
 			$this->db->query("DELETE FROM station_profile WHERE user_id = ?",$user_id);
 			$this->db->query("DELETE FROM station_logbooks WHERE user_id = ?",$user_id);
-			$this->db->query("delete from user_options where user_id=?",$user_id);
+			$this->db->query("DELETE FROM user_options WHERE user_id=?",$user_id);
 			$this->db->query("DELETE FROM ".$this->config->item('auth_table')." WHERE user_id = ?",$user_id);
 			return 1;
 		} else {
@@ -428,7 +453,7 @@ class User_Model extends CI_Model {
 	// FUNCTION: void update_session()
 	// Updates a user's login session after they've logged in
 	// TODO: This should return bool TRUE/FALSE or 0/1
-	function update_session($id, $u = null, $impersonate = false) {
+	function update_session($id, $u = null, $impersonate = false, $custom_data = null) {
 
 		if ($u == null) {
 			$u = $this->get_by_id($id);
@@ -475,7 +500,13 @@ class User_Model extends CI_Model {
 			'isWinkeyEnabled' => $u->row()->winkey,
 			'hasQrzKey' => $this->hasQrzKey($u->row()->user_id),
 			'impersonate' => $this->session->userdata('impersonate') ?? false,
+			'clubstation' => $u->row()->clubstation,
+			'source_uid' => $this->session->userdata('source_uid') ?? ''
 		);
+
+		if ($this->config->item('special_callsign')) {
+			$userdata['available_clubstations'] = $this->get_clubstations($u->row()->user_id) ?? 'none';
+		}
 
 		foreach (array_keys($this->frequency->defaultFrequencies) as $band) {
 			$qrg_unit = $this->session->userdata("qrgunit_$band") ?? ($this->user_options_model->get_options('frequency', array('option_name' => 'unit', 'option_key' => $band), $u->row()->user_id)->row()->option_value ?? '');
@@ -486,8 +517,25 @@ class User_Model extends CI_Model {
 			}
 		}
 
+		// Restore custom data in impersonation mode
+		foreach ($this->session->userdata() as $key => $value) {
+			if (substr($key, 0, 3) == 'cd_') {
+				$userdata[$key] = $value;
+			}
+		}
+
+		// Overrides
 		if ($impersonate) {
 			$userdata['impersonate'] = true;
+			$userdata['available_clubstations'] = $this->get_clubstations($u->row()->user_id);
+		}
+		if ($userdata['clubstation'] == 1) {
+			$userdata['available_clubstations'] = 'none';
+		}
+		if (isset($custom_data)) {
+			foreach ($custom_data as $key => $value) {
+				$userdata['cd_' . $key] = $value;
+			}
 		}
 
 		$this->session->set_userdata($userdata);
@@ -502,6 +550,7 @@ class User_Model extends CI_Model {
 		{
 			$user_id = $this->session->userdata('user_id');
 			$user_type = $this->session->userdata('user_type');
+			$src_user_type = $this->session->userdata('cd_src_user_type');
 			$user_hash = $this->session->userdata('user_hash');
 			$impersonate = $this->session->userdata('impersonate');
 
@@ -515,7 +564,7 @@ class User_Model extends CI_Model {
 					return 0;
 				}
 			} else {  // handle the maintenance mode and kick out user on page reload if not an admin
-				if($user_type == '99' || $impersonate === true) {
+				if($user_type == '99' || $src_user_type === '99') {
 					if($this->_auth($user_id."-".$user_type, $user_hash)) {
 						// Freshen the session
 						$this->update_session($user_id, $u);
@@ -538,8 +587,14 @@ class User_Model extends CI_Model {
 	// Authenticate a user against the users table
 	function authenticate($username, $password) {
 		$u = $this->get($username);
-		if($u->num_rows() != 0)
-		{
+		if($u->num_rows() != 0) {
+			// direct login to clubstations are not allowed
+			if ($u->row()->clubstation == 1) {
+				$uid = $u->row()->user_id;
+				log_message('debug', "User ID: [$uid] Login rejected because of a external clubstation login attempt.");
+				return 2;
+			}
+
 			if($this->_auth($password, $u->row()->user_password)) {
 				if (ENVIRONMENT != "maintenance") {
 					return 1;
@@ -582,7 +637,7 @@ class User_Model extends CI_Model {
 
 				$this->set_last_seen($u->row()->user_id);
 			}
-			return 1;
+				return 1;
 		} else {
 			return 0;
 		}
@@ -590,20 +645,41 @@ class User_Model extends CI_Model {
 
 	// FUNCTION: object users()
 	// Returns a list of users with additional counts
-	function users() {
-		$this->db->select('(SELECT COUNT(*) FROM station_profile WHERE user_id = users.user_id) as stationcount');
-		$this->db->select('(SELECT COUNT(*) FROM station_logbooks WHERE user_id = users.user_id) as logbookcount');
-		$this->db->select('(SELECT COUNT(*) FROM ' . $this->config->item('table_name') . ' WHERE station_id IN (SELECT station_id from station_profile WHERE user_id = users.user_id)) as qsocount');
-		$this->db->select('
-			(SELECT COUNT(*) FROM ' . $this->config->item('table_name') . ' WHERE station_id IN (SELECT station_id FROM station_profile WHERE user_id = users.user_id)) as qsocount,
-			(SELECT MAX(COL_TIME_ON) FROM ' . $this->config->item('table_name') . ' WHERE station_id IN (SELECT station_id FROM station_profile WHERE user_id = users.user_id)) as lastqso
-		');
-		$this->db->select('users.*');
-		$this->db->from('users');
-
-		$result = $this->db->get();
-
+	function users($club = '') {
+		$sql="  SELECT COUNT(distinct sp.station_id) AS stationcount, count(distinct sl.logbook_id) AS logbookcount, count(distinct log.col_primary_key) AS qsocount,
+			MAX(COL_TIME_ON) AS lastqso,
+			u.*
+			FROM  users u
+			LEFT OUTER JOIN station_profile sp ON (sp.user_id = u.user_id)
+			LEFT OUTER JOIN station_logbooks sl ON (sl.user_id = u.user_id)
+			LEFT OUTER JOIN ". $this->config->item('table_name') ." log on (log.station_id=sp.station_id)";
+		if ($this->config->item('special_callsign')) {
+			if ($club == 'is_club') {
+				$sql.=' WHERE clubstation=1';
+			} else {
+				$sql.=' WHERE clubstation!=1';
+			}
+		}
+		$sql.=" GROUP BY u.user_id";
+		$result = $this->db->query($sql);
+		if ($this->config->item('special_callsign')) {
+			if ($club == 'is_club') {
+				foreach ($result->result() as &$row) {
+					$row->lastoperator=$this->get_last_op($row->user_id,$row->lastqso);
+				}
+			} else {
+				foreach ($result->result() as &$row) {
+					$row->lastoperator='';
+				}
+			}
+		}
 		return $result;
+	}
+
+	function get_last_op($userid,$lastqso) {
+		$sql="SELECT log.COL_OPERATOR FROM ". $this->config->item('table_name') ." log INNER JOIN station_profile sp ON (log.station_id=sp.station_id) where sp.user_id=? AND col_time_on=? ORDER BY col_time_on DESC LIMIT 1";
+		$resu=$this->db->query($sql,array($userid,$lastqso));
+		return $resu->result()[0]->COL_OPERATOR ?? '';
 	}
 
 	// FUNCTION: array timezones()
@@ -739,6 +815,38 @@ class User_Model extends CI_Model {
 		} else {
 			return false;
 		}
+	}
+
+	function get_clubstations($user_id) {
+		$this->load->model('club_model');
+		$clubstations = $this->club_model->get_clubstations($user_id);
+
+		return $clubstations;
+	}
+
+	function convert($user_id, $clubstation) {
+		$clubstation_value = ($clubstation == true) ? 1 : 0;
+	
+		$sql = "UPDATE users SET clubstation = ? WHERE user_id = ?;";
+	
+		$this->db->trans_start();
+	
+		if (!$this->db->query($sql, [$clubstation_value, $user_id])) {
+			$this->db->trans_rollback();
+			return false;
+		}
+	
+		if ($clubstation) {
+			$delete_sql = "DELETE FROM club_permissions WHERE club_id = ?;";
+			if (!$this->db->query($delete_sql, [$user_id])) {
+				$this->db->trans_rollback();
+				return false;
+			}
+		}
+	
+		$this->db->trans_complete();
+	
+		return $this->db->trans_status();
 	}
 
 }
